First push
Home rental is an everyday business in the ever busy Lagos. This app helps a prospective tenant to have a fair idea of the homes in different areas of Lagos as at May 2024. A landlord or a property developer can also use the app to predict their rental incomes. In developing the app, the features and rental value of data of over 21000 Lagos properties were scraped from the website of a popular online property agent in Nigeria. Please note that the app is for academic purposes only. Areas of improvement for the project will include: 1). There is need to get data over a period of years so that the effect of time and inflation can be built into the model. 2)Other home features such as (condition of the property, either newly built or old house) are not stated on the scraped website. This can be included into the model because it has a significant effect on rental value